Adjust for keyboard notification. File structure update. Minor updates
